refactor(core): extract duplicated utils to @docubook/core
- Add clsx and tailwind-merge deps to @docubook/core
- Create shared utils in packages/core/src/utils.ts
- Re-export from apps/web/lib/utils.ts
- Re-export from packages/flame/.docu/lib/utils.ts

Refs: #125
