+++ packages/core/package.json
@@ -41,12 +41,14 @@
   "license": "MIT",
   "dependencies": {
     "@11ty/gray-matter": "^2.0.2",
+    "clsx": "^2.1.1",
     "next-mdx-remote": "^6.0.0",
     "rehype-autolink-headings": "^7.1.0",
     "rehype-code-titles": "^1.2.1",
     "rehype-prism-plus": "^2.0.2",
     "rehype-slug": "^6.0.0",
     "remark-gfm": "^4.0.1",
+    "tailwind-merge": "^2.6.1",
     "unist-util-visit": "^5.1.0"
   },
   "devDependencies": {
